chore: extract inline benches to a separate directory
