+++ .cargo/config.toml
@@ -1,6 +1,3 @@
-[alias]
-criterion = "bench --bench=criterion --features=bench"
-
 # This enables KaTex in docs, but requires running `cargo doc --no-deps`.
 [build]
 rustdocflags = "--html-in-header .cargo/katex-header.html"
