+++ .github/workflows/ci.yml
@@ -93,8 +93,8 @@ jobs:
         uses: taiki-e/install-action@cargo-llvm-cov
       - name: Build with code coverage
         run: |
-          cargo llvm-cov --no-report --locked --workspace --all-features --all-targets -- --help
-          cargo llvm-cov --no-report --locked --workspace --all-features --doc -- --help
+          cargo llvm-cov --no-report --locked --workspace --all-features --all-targets
+          cargo llvm-cov --no-report --locked --workspace --all-features --doc
       - name: Generate code coverage
         env:
           # Covered tests are slower, so reduce the test load. (We also shouldn't
