+++ src/algorithms/mul.rs
@@ -99,7 +99,7 @@ pub fn addmul(mut lhs: &mut [u64], mut a: &[u64], mut b: &[u64]) -> bool {
 
     // Iterate over limbs of `b` and add partial products to `lhs`.
     let mut overflow = false;
-    for &b in b.iter() {
+    for &b in b {
         if lhs.len() >= a.len() {
             let (target, rest) = lhs.split_at_mut(a.len());
             let carry = addmul_nx1(target, a, b);
@@ -123,7 +123,7 @@ pub fn add_nx1(lhs: &mut [u64], mut a: u64) -> u64 {
     if a == 0 {
         return 0;
     }
-    for lhs in lhs.iter_mut() {
+    for lhs in lhs {
         let sum = u128::add(*lhs, a);
         *lhs = sum.low();
         a = sum.high();
@@ -340,39 +340,5 @@ mod tests {
             8411211985208067428
         ]);
         assert_eq!(borrow, 17196576577663999042);
-    }
-}
-
-#[cfg(feature = "bench")]
-#[doc(hidden)]
-pub mod bench {
-    use crate::const_for;
-
-    use super::*;
-    use criterion::{black_box, BatchSize, Criterion};
-    use rand::Rng;
-
-    pub fn group(criterion: &mut Criterion) {
-        bench_addmul_nnn(criterion);
-    }
-
-    fn bench_addmul_nnn(criterion: &mut Criterion) {
-        const_for!(SIZE in [0,1,2,3,4,5,6] {
-            let mut rng = rand::thread_rng();
-            criterion.bench_function(&format!("algo/addmul_n/{SIZE}"), move |bencher| {
-                bencher.iter_batched(
-                    || (
-                        rng.gen::<[u64; SIZE]>(),
-                        rng.gen::<[u64; SIZE]>(),
-                        rng.gen::<[u64; SIZE]>(),
-                    ),
-                    |(mut lhs, a, b)| {
-                        addmul_n(&mut lhs, &a, &b);
-                        black_box(lhs)
-                    },
-                    BatchSize::SmallInput,
-                );
-            });
-        });
     }
 }
