+++ src/const_for.rs
@@ -49,17 +49,17 @@ macro_rules! const_for {
         })*
     };
     ($C:ident in SIZES $x:block) => {
-        const_for!($C in [0] $x);
-        const_for!($C in NON_ZERO $x);
+        $crate::const_for!($C in [0] $x);
+        $crate::const_for!($C in NON_ZERO $x);
     };
     ($C:ident in NON_ZERO $x:block) => {
-        const_for!($C in [1, 2, 63, 64, 65, 127, 128, 129, 256, 384, 512, 4096] $x);
+        $crate::const_for!($C in [1, 2, 63, 64, 65, 127, 128, 129, 256, 384, 512, 4096] $x);
     };
     ($C:ident in BENCH $x:block) => {
-        const_for!($C in [0, 64, 128, 192, 256, 384, 512, 4096] $x);
+        $crate::const_for!($C in [0, 64, 128, 192, 256, 384, 512, 4096] $x);
     };
     ($C:ident in $S:ident if ( $c:expr ) $x:block) => {
-        const_for!($C in $S {
+        $crate::const_for!($C in $S {
             if $c {
                 $x
             }
