+++ src/support/quickcheck.rs
@@ -9,7 +9,7 @@ impl<const BITS: usize, const LIMBS: usize> Arbitrary for Uint<BITS, LIMBS> {
     fn arbitrary(g: &mut Gen) -> Self {
         let mut limbs = [0; LIMBS];
         if let Some((last, rest)) = limbs.split_last_mut() {
-            for limb in rest.iter_mut() {
+            for limb in rest {
                 *limb = u64::arbitrary(g);
             }
             *last = u64::arbitrary(g) & Self::MASK;
