+++ src/support/rand.rs
@@ -14,7 +14,7 @@ impl<const BITS: usize, const LIMBS: usize> Distribution<Uint<BITS, LIMBS>> for 
     fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> Uint<BITS, LIMBS> {
         let mut limbs = [0; LIMBS];
         if let Some((last, rest)) = limbs.split_last_mut() {
-            for limb in rest.iter_mut() {
+            for limb in rest {
                 *limb = rng.gen();
             }
             *last = Uniform::new_inclusive(0, Uint::<BITS, LIMBS>::MASK).sample(rng);
